Track memory bytes per container
We have a problem where the gauge without labels is created and always
set to zero. I don't want this to be there but haven't figured out how
to get rid of it yet.

The answer might have to be that we change the prometheus library we're
using. This one has not had updates in a while.